+++ example/ValidatorChainExample.php
@@ -3,16 +3,15 @@
 require __DIR__.'/../vendor/autoload.php';
 
 
+use LDL\Validators\Chain\Dumper\ValidatorChainJsonDumper;
+use LDL\Validators\Chain\Dumper\ValidatorChainPhpDumper;
+use LDL\Validators\NumericComparisonValidator;
 use LDL\Validators\RegexValidator;
 use LDL\Validators\Chain\AndValidatorChain;
 use LDL\Validators\Chain\OrValidatorChain;
 use LDL\Validators\IntegerValidator;
 use LDL\Validators\Chain\Exception\CombinedException;
 use LDL\Validators\StringValidator;
-use LDL\Validators\MaxNumericValidator;
-use LDL\Validators\MinNumericValidator;
-use LDL\Validators\Chain\Dumper\ValidatorChainPhpDumper;
-use LDL\Validators\Chain\Dumper\ValidatorChainJsonDumper;
 use LDL\Validators\Chain\Dumper\ValidatorChainExprDumper;
 
 echo "Create Validator Chain\n";
@@ -27,19 +26,19 @@ $chain = new OrValidatorChain([
         new OrValidatorChain([
             new IntegerValidator(),
             new AndValidatorChain([
-                new MaxNumericValidator(500),
-                new MinNumericValidator(10),
+                new NumericComparisonValidator(500, '>'),
+                new NumericComparisonValidator(10, '<='),
             ])
         ])
     ]),
     new OrValidatorChain([
         new RegexValidator('#[a-z]+#')
     ])
-],true);
+],false);
 
 echo "Validate: abc\n";
 
-$chain->validate('@abc');
+$chain->validate('abc');
 
 echo "Validate: 123\n";
 $chain->validate(123);
@@ -50,16 +49,16 @@ try{
     dump("EXCEPTION: {$e->getCombinedMessage()}");
 }
 
-echo "Dump chain as boolean expression:\n\n";
+echo "\nDump chain as boolean expression:\n";
 echo ValidatorChainExprDumper::dump($chain);
 
-echo "Dump chain as PHP:\n\n";\
-sleep(5);
+echo "\nDump chain as PHP:\n";
+sleep(3);
 dump(ValidatorChainPhpDumper::dump($chain));
 
 
-echo "Dump chain as JSON:\n\n";
-sleep(5);
+echo "\nDump chain as JSON:\n";
+sleep(3);
 echo ValidatorChainJsonDumper::dump($chain)."\n\n";
 
-\LDL\Validators\Chain\Loader\ValidatorChainJsonLoader::load(ValidatorChainJsonDumper::dump($chain));
+//\LDL\Validators\Chain\Loader\ValidatorChainJsonLoader::load(ValidatorChainJsonDumper::dump($chain));